backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
paddlepaddle>=3.0.0
paddleocr>=2.7.0
tensorflow==2.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0

